OV-114: + verify user
